+++ src/frontend/org/voltdb/sysprocs/Statistics.java
@@ -249,7 +249,7 @@ public class Statistics extends VoltSystemProcedure {
             // Choose the lowest site ID on this host to do the scan
             // All other sites should just return empty results tables.
             int host_id = context.getExecutionSite().getCorrespondingHostId();
-            Integer lowest_site_id =
+            Long lowest_site_id =
                 VoltDB.instance().getCatalogContext().siteTracker.
                 getLowestLiveExecSiteIdForHost(host_id);
             if (context.getExecutionSite().getSiteId() != lowest_site_id)
@@ -279,7 +279,7 @@ public class Statistics extends VoltSystemProcedure {
             // Choose the lowest site ID on this host to do the scan
             // All other sites should just return empty results tables.
             int hostId = context.getExecutionSite().getCorrespondingHostId();
-            Integer lowestSiteId =
+            Long lowestSiteId =
                 VoltDB.instance().getCatalogContext().siteTracker.
                 getLowestLiveExecSiteIdForHost(hostId);
             if (context.getExecutionSite().getSiteId() == lowestSiteId) {
@@ -306,7 +306,7 @@ public class Statistics extends VoltSystemProcedure {
             // Choose the lowest site ID on this host to do the scan
             // All other sites should just return empty results tables.
             int host_id = context.getExecutionSite().getCorrespondingHostId();
-            Integer lowest_site_id =
+            Long lowest_site_id =
                 VoltDB.instance().getCatalogContext().siteTracker.
                 getLowestLiveExecSiteIdForHost(host_id);
             if (context.getExecutionSite().getSiteId() == lowest_site_id)
